Speedup video rendering

+++ render.py
@@ -11,6 +11,7 @@ import numpy as np
 import lightning
 import torch
 import torchvision
+import mediapy
 from tqdm import tqdm
 from internal.cameras.cameras import Cameras
 from internal.renderers.vanilla_renderer import VanillaRenderer
@@ -102,39 +103,83 @@ def parse_model_transformations(camera_path: dict) -> list[list]:
     return frame_transformation_list
 
 
-def save_image(image_information: tuple):
-    image, output_path = image_information
+def save_image(image, output_path):
     torchvision.utils.save_image(image, output_path)
 
 
-def process_image_queue(image_queue: queue.Queue):
+def process_save_image_queue(image_queue: queue.Queue, frame_output_path: str):
     while True:
         image_information = image_queue.get()
         if image_information is None:
             break
         try:
-            save_image(image_information)
+            save_image(image_information[0], os.path.join(frame_output_path, "{:06d}.png".format(image_information[1])))
         except:
             traceback.print_exc()
+
+
+def process_image_to_video_queue(image_queue: queue.Queue, video_writer: mediapy.VideoWriter):
+    while True:
+        image_information = image_queue.get()
+        if image_information is None:
+            break
+        video_writer.add_image(image_information[0].permute(1, 2, 0).numpy())
+
+
+def process_image_queue(image_save_batch: int, image_queue: queue.Queue, video_writer: mediapy.VideoWriter, frame_output_path: str):
+    class MockQueue:
+        def put(self, *args, **kwargs):
+            return
+
+    # create image to file threads
+    save_image_threads = []
+    frame_saving_queue = MockQueue()
+    if frame_output_path is not None:
+        frame_saving_queue = queue.Queue(maxsize=max(image_save_batch // 2, 1))
+        for _ in range(image_save_batch):
+            thread = threading.Thread(target=process_save_image_queue, args=(frame_saving_queue, frame_output_path))
+            save_image_threads.append(thread)
+            thread.start()
+
+    # create image to video thread
+    image_to_video_queue = queue.Queue(maxsize=1)
+    image_to_video_thread = threading.Thread(target=process_image_to_video_queue, args=(image_to_video_queue, video_writer))
+    image_to_video_thread.start()
+
+    # forward rendered image to threads
+    while True:
+        image_information = image_queue.get()
+        if image_information is None:
+            break
+        frame_saving_queue.put(image_information)
+        image_to_video_queue.put(image_information)
+
+    # wait for all threads finishing
+    for _ in range(len(save_image_threads)):
+        frame_saving_queue.put(None)
+    for i in save_image_threads:
+        i.join()
+    image_to_video_queue.put(None)
+    image_to_video_thread.join()
 
 
 def render_frames(
         cameras: Cameras,
         model_transformations: list,
         viewer_renderer: ViewerRenderer,
-        output_path: str,
+        frame_output_path: str,
+        video_writer: mediapy.VideoWriter,
         image_save_batch: int,
         device,
 ):
-    os.makedirs(output_path, exist_ok=True)
-
-    # create queue and threads
-    image_queue = queue.Queue(maxsize=max(image_save_batch // 2, 1))
-    threads = []
-    for _ in range(image_save_batch):
-        thread = threading.Thread(target=process_image_queue, args=(image_queue,))
-        threads.append(thread)
-        thread.start()
+    image_queue = queue.Queue(maxsize=1)
+    image_process_thread = threading.Thread(target=process_image_queue, args=(
+        image_save_batch,
+        image_queue,
+        video_writer,
+        frame_output_path,
+    ))
+    image_process_thread.start()
 
     for idx in tqdm(range(len(cameras)), desc="rendering frames"):
         # model transform
@@ -149,15 +194,10 @@ def render_frames(
         # render
         camera = cameras[idx].to_device(device)
         image = viewer_renderer.get_outputs(camera).cpu()
-        image_output_path = os.path.join(output_path, "{:06d}.png".format(idx))
+        image_queue.put((image, idx))
 
-        image_queue.put((image, image_output_path))
-
-    # wait for all threads finishing image saving
-    for _ in range(len(threads)):
-        image_queue.put(None)
-    for i in threads:
-        i.join()
+    image_queue.put(None)
+    image_process_thread.join()
 
 
 if __name__ == "__main__":
@@ -165,6 +205,8 @@ if __name__ == "__main__":
     parser.add_argument("model_paths", type=str, nargs="+")
     parser.add_argument("--camera-path-filename", type=str, required=True)
     parser.add_argument("--output-path", type=str, required=True)
+    parser.add_argument("--save-images", "--save-image", "--save_image", "--save-frames", action="store_true",
+                        help="Whether save each frame to an image file")
     parser.add_argument("--image-save-batch", "-b", type=int, default=8,
                         help="increase this to speedup rendering, but more memory will be consumed")
     parser.add_argument("--disable-transform", action="store_true", default=False)
@@ -176,11 +218,14 @@ if __name__ == "__main__":
     with open(args.camera_path_filename, "r") as f:
         camera_path = json.load(f)
 
+    # whether a 2DGS model
     renderer_override = None
     if args.vanilla_gs2d is True:
         from internal.renderers.vanilla_2dgs_renderer import Vanilla2DGSRenderer
 
         renderer_override = Vanilla2DGSRenderer()
+
+    # instantiate renderer
     renderer = initializer_viewer_renderer(
         args.model_paths,
         enable_transform=camera_path["enable_transform"],
@@ -190,37 +235,38 @@ if __name__ == "__main__":
         device=device,
     )
 
+    # load cameras
     cameras = parse_camera_poses(camera_path)
     if args.disable_transform is False:
         model_transformations = parse_model_transformations(camera_path)
     else:
         model_transformations = [[] for _ in range(len(cameras))]
 
-    frame_output_path = args.output_path + "_frames"
-    for i in glob.glob(os.path.join(frame_output_path, "*.png")):
-        os.unlink(i)
-    with torch.no_grad():
+    # create output path
+    os.makedirs(os.path.dirname(args.output_path), exist_ok=True)
+    frame_output_path = None
+    if args.save_images is True:
+        frame_output_path = args.output_path + "_frames"
+        os.makedirs(frame_output_path, exist_ok=True)
+        for i in glob.glob(os.path.join(frame_output_path, "*.png")):
+            os.unlink(i)
+
+    # start rendering
+    with torch.no_grad(), mediapy.VideoWriter(
+            path=args.output_path,
+            shape=(cameras[0].height.item(), cameras[0].width.item()),
+            fps=camera_path["fps"],
+    ) as video_writer:
         render_frames(
             cameras,
             model_transformations,
             viewer_renderer=renderer,
-            output_path=frame_output_path,
+            frame_output_path=frame_output_path,
+            video_writer=video_writer,
             image_save_batch=args.image_save_batch,
             device=device,
         )
 
-    subprocess.call([
-        "ffmpeg",
-        "-y",
-        "-framerate", str(camera_path["fps"]),
-        "-i", os.path.join(frame_output_path, "%06d.png"),
-        "-pix_fmt", "yuv420p",
-        args.output_path,
-    ])
-
-    try:
-        subprocess.call(["stty", "sane"])
-    except:
-        pass
-
-    print(f"Video saved to `{args.output_path}`")
+    if frame_output_path is not None:
+        print(f"Video frames saved to '{frame_output_path}'")
+    print(f"Video saved to '{args.output_path}'")
